backend/requirements.txt

praw>=7.7.1
pandas>=2.0.0
vaderSentiment>=3.3.2
fastapi>=0.95.0
uvicorn>=0.21.1
python-dotenv>=1.0.0
nltk>=3.8.1
wordcloud>=1.9.2
matplotlib>=3.7.1
scikit-learn>=1.2.2
python-multipart>=0.0.6
httpx>=0.24.1
python-dateutil>=2.8.2
urllib3>=1.26.0
GoogleNews